+++ src/Data/Profunctor/Distributor.hs
@@ -8,31 +8,25 @@ Stability   : provisional
 Portability : non-portable
 -}
 
-{-# OPTIONS_GHC -Wno-orphans #-}
-
 module Data.Profunctor.Distributor
-  ( -- * Monoidal
-    Monoidal, oneP, (>*<), (>*), (*<), dimap2, foreverP, replicateP
-  , meander, (>:<), asEmpty
-    -- * Distributor
-  , Distributor (..), dialt, Homogeneous (homogeneously)
-    -- * Alternator/Filtrator
-  , Alternator (..), Filtrator (filtrate)
+  ( -- * Distributor
+    Distributor (..), dialt
+    -- * Alternator
+  , Alternator (..)
+    -- * Homogeneous
+  , Homogeneous (..)
   ) where
 
 import Control.Applicative hiding (WrappedArrow)
 import Control.Applicative qualified as Ap (WrappedArrow)
 import Control.Arrow
 import Control.Lens hiding (chosen)
-import Control.Lens.Internal.Context
 import Control.Lens.Internal.Profunctor
 import Control.Lens.PartialIso
-import Control.Monad
 import Data.Bifunctor.Clown
 import Data.Bifunctor.Joker
 import Data.Bifunctor.Product
 import Data.Complex
-import Data.Distributive
 import Data.Functor.Adjunction
 import Data.Functor.Compose
 import Data.Functor.Contravariant.Divisible
@@ -44,6 +38,7 @@ import Data.Profunctor qualified as Pro (WrappedArrow)
 import Data.Profunctor.Cayley
 import Data.Profunctor.Composition
 import Data.Profunctor.Monad
+import Data.Profunctor.Monoidal
 import Data.Profunctor.Yoneda
 import Data.Proxy
 import Data.Sequence (Seq)
@@ -52,113 +47,6 @@ import Data.Tree (Tree (..))
 import Data.Vector (Vector)
 import Data.Void
 import GHC.Generics
-import Witherable
-
--- Monoidal --
-
-{- | A lax `Monoidal` product `Profunctor` has unit `oneP`
-and product `>*<` lax monoidal structure morphisms.
-This is equivalent to the `Profunctor` also being `Applicative`.
-
-Laws:
-
->>> let (f >< g) (a,c) = (f a, g c)
->>> let lunit = dimap (\((),a) -> a) (\a -> ((),a))
->>> let runit = dimap (\(a,()) -> a) (\a -> (a,()))
->>> let assoc = dimap (\(a,(b,c)) -> ((a,b),c)) (\((a,b),c) -> (a,(b,c)))
-
-prop> dimap (f >< g) (h >< i) (p >*< q) = dimap f h p >*< dimap g i q
-prop> oneP >*< p = lunit p
-prop> p >*< oneP = runit p
-prop> p >*< q >*< r = assoc ((p >*< q) >*< r)
-
--}
-type Monoidal p = (Profunctor p, forall x. Applicative (p x))
-
-{- | `oneP` is the unit of a `Monoidal` `Profunctor`. -}
-oneP :: Monoidal p => p () ()
-oneP = pure ()
-
-{- | `>*<` is the product of a `Monoidal` `Profunctor`. -}
-(>*<) :: Monoidal p => p a b -> p c d -> p (a,c) (b,d)
-(>*<) = dimap2 fst snd (,)
-infixr 6 >*<
-
-{- | `>*` sequences actions, discarding the value of the first argument;
-analagous to `*>`, extending it to `Monoidal`.
-
-prop> oneP >* p = p
-
--}
-(>*) :: Monoidal p => p () c -> p a b -> p a b
-x >* y = lmap (const ()) x *> y
-infixl 5 >*
-
-{- | `*<` sequences actions, discarding the value of the second argument;
-analagous to `<*`, extending it to `Monoidal`.
-
-prop> p *< oneP = p
-
--}
-(*<) :: Monoidal p => p a b -> p () c -> p a b
-x *< y = x <* lmap (const ()) y
-infixl 5 *<
-
-{- | `dimap2` is a curried, functionalized form of `>*<`,
-analagous to `liftA2`. -}
-dimap2
-  :: Monoidal p
-  => (s -> a)
-  -> (s -> c)
-  -> (b -> d -> t)
-  -> p a b -> p c d -> p s t
-dimap2 f g h p q = liftA2 h (lmap f p) (lmap g q)
-
-{- | `foreverP` repeats an action indefinitely;
-analagous to `forever`, extending it to `Monoidal`. -}
-foreverP :: Monoidal p => p () c -> p a b
-foreverP a = let a' = a >* a' in a'
-
-{- | Thanks to Fy on Monoidal Café Discord.
-
-`replicateP` is roughly analagous to `replicateM`,
-repeating an action a number of times.
-However, instead of an `Int` term, it expects
-a `Traversable` & `Distributive` type. Such a
-type is a homogeneous countable product.
--}
-replicateP
-  :: (Traversable t, Distributive t, Monoidal p)
-  => p a b -> p (t a) (t b)
-replicateP p = traverse (\f -> lmap f p) (distribute id)
-
-{- | For any `Monoidal`, `Choice` & `Strong` `Profunctor`,
-`meander` is invertible and gives a default implementation for the
-`Data.Profunctor.Traversing.wander`
-method of `Data.Profunctor.Traversing.Traversing`,
-though `Strong` is not needed for its definition.
-
-See Pickering, Gibbons & Wu,
-[Profunctor Optics - Modular Data Accessors](https://arxiv.org/abs/1703.10857)
--}
-meander
-  :: (Monoidal p, Choice p)
-  => ATraversal s t a b -> p a b -> p s t
-meander f = dimap (f sell) iextract . trav
-  where
-    trav
-      :: (Monoidal q, Choice q)
-      => q u v -> q (Bazaar (->) u w x) (Bazaar (->) v w x)
-    trav q = mapIso funListEot $ right' (q >*< trav q)
-
-{- | A `Monoidal` nil operator. -}
-asEmpty :: (AsEmpty s, Monoidal p, Choice p) => p s s
-asEmpty = _Empty >? oneP
-
-{- | A `Monoidal` cons operator. -}
-(>:<) :: (Cons s t a b, Monoidal p, Choice p) => p a b -> p s t -> p s t
-x >:< xs = _Cons >? x >*< xs
-infixr 5 >:<
 
 -- Distributor --
 
@@ -216,11 +104,11 @@ class Monoidal p => Distributor p where
 
   {- | One or none. -}
   optionalP :: p a b -> p (Maybe a) (Maybe b)
-  optionalP p = mapIso maybeEot (oneP >+< p)
+  optionalP p = mapIso eotMaybe (oneP >+< p)
 
   {- | Zero or more. -}
   manyP :: p a b -> p [a] [b]
-  manyP p = mapIso listEot (oneP >+< p >*< manyP p)
+  manyP p = mapIso eotList (oneP >+< p >*< manyP p)
 
 instance Distributor (->) where
   zeroP = id
@@ -403,15 +291,15 @@ instance Homogeneous Maybe where
 instance Homogeneous [] where
   homogeneously = manyP
 instance Homogeneous Vector where
-  homogeneously p = mapIso listEot (oneP >+< p >*< homogeneously p)
+  homogeneously p = mapIso eotList (oneP >+< p >*< homogeneously p)
 instance Homogeneous Seq where
-  homogeneously p = mapIso listEot (oneP >+< p >*< homogeneously p)
+  homogeneously p = mapIso eotList (oneP >+< p >*< homogeneously p)
 instance Homogeneous Complex where
   homogeneously p = dimap2 realPart imagPart (:+) p p
 instance Homogeneous Tree where
   homogeneously p = dimap2 rootLabel subForest Node p (manyP (homogeneously p))
 
--- Alternator/Filtrator --
+-- Alternator --
 
 {- | The `Alternator` class co-extends `Choice` and `Distributor`,
 as well as `Alternative`, adding the `alternate` method,
@@ -474,159 +362,3 @@ instance Alternator p => Alternator (Yoneda p) where
   alternate (Left p) = proreturn (alternate (Left (proextract p)))
   alternate (Right p) = proreturn (alternate (Right (proextract p)))
   someP = proreturn . someP . proextract
-
-{- | The `Filtrator` class extends `Cochoice`,
-as well as `Filterable`, adding the `filtrate` method,
-which is an oplax monoidal structure morphism dual to `>+<`.
--}
-class (Cochoice p, forall x. Filterable (p x))
-  => Filtrator p where
-
-    {- |
-    prop> unleft = fst . filtrate
-    prop> unright = snd . filtrate
-
-    `filtrate` is a distant relative to `Data.Either.partitionEithers`.
-
-    `filtrate` has a default for `Choice`.
-    -}
-    filtrate
-      :: p (Either a c) (Either b d)
-      -> (p a b, p c d)
-    default filtrate
-      :: Choice p
-      => p (Either a c) (Either b d)
-      -> (p a b, p c d)
-    filtrate =
-      dimapMaybe (Just . Left) (either Just (pure Nothing))
-      &&&
-      dimapMaybe (Just . Right) (either (pure Nothing) Just)
-
-instance (Profunctor p, forall x. Functor (p x), Filterable f)
-  => Filtrator (WrappedPafb f p) where
-    filtrate (WrapPafb p) =
-      ( WrapPafb $ dimap Left (mapMaybe (either Just (const Nothing))) p
-      , WrapPafb $ dimap Right (mapMaybe (either (const Nothing) Just)) p
-      )
-instance Filtrator p => Filtrator (Coyoneda p) where
-  filtrate p =
-    let (q,r) = filtrate (proextract p)
-    in (proreturn q, proreturn r)
-instance Filtrator p => Filtrator (Yoneda p) where
-  filtrate p =
-    let (q,r) = filtrate (proextract p)
-    in (proreturn q, proreturn r)
-instance Filtrator (Forget r) where
-  filtrate (Forget f) = (Forget (f . Left), Forget (f . Right))
-instance (Filterable f, Traversable f) => Filtrator (Star f) where
-  filtrate (Star f) =
-    ( Star (mapMaybe (either Just (const Nothing)) . f . Left)
-    , Star (mapMaybe (either (const Nothing) Just) . f . Right)
-    )
-instance Filtrator (PartialExchange a b) where
-  filtrate (PartialExchange f g) =
-    ( PartialExchange (f . Left) (either Just (pure Nothing) <=< g)
-    , PartialExchange (f . Right) (either (pure Nothing) Just <=< g)
-    )
-
--- FunList --
-
-{- |
-`FunList` is isomorphic to `Bazaar` @(->)@.
-It's needed to define `meander`.
-
-See van Laarhoven, A non-regular data type challenge
-[https://twanvl.nl/blog/haskell/non-regular1]
--}
-data FunList a b t
-  = DoneFun t
-  | MoreFun a (Bazaar (->) a b (b -> t))
-instance Functor (FunList a b) where
-  fmap f = \case
-    DoneFun t -> DoneFun (f t)
-    MoreFun a h -> MoreFun a (fmap (f .) h)
-instance Applicative (FunList a b) where
-  pure = DoneFun
-  (<*>) = \case
-    DoneFun t -> fmap t
-    MoreFun a h -> \l ->
-      MoreFun a (flip <$> h <*> fromFun l)
-instance Sellable (->) FunList where sell b = MoreFun b (pure id)
-
-toFun :: Bazaar (->) a b t -> FunList a b t
-toFun (Bazaar f) = f sell
-
-fromFun :: FunList a b t -> Bazaar (->) a b t
-fromFun = \case
-  DoneFun t -> pure t
-  MoreFun a f -> ($) <$> f <*> sell a
-
-funListEot :: Iso
-  (Bazaar (->) a1 b1 t1) (Bazaar (->) a2 b2 t2)
-  (Either t1 (a1, Bazaar (->) a1 b1 (b1 -> t1)))
-  (Either t2 (a2, Bazaar (->) a2 b2 (b2 -> t2)))
-funListEot = iso toFun fromFun . iso f g where
-  f = \case
-    DoneFun t -> Left t
-    MoreFun a baz -> Right (a, baz)
-  g = \case
-    Left t -> DoneFun t
-    Right (a, baz) -> MoreFun a baz
-
--- Orphanage --
-
-instance Monoid r => Applicative (Forget r a) where
-  pure _ = Forget mempty
-  Forget f <*> Forget g = Forget (f <> g)
-instance Decidable f => Applicative (Clown f a) where
-  pure _ = Clown conquer
-  Clown x <*> Clown y = Clown (divide (id &&& id) x y)
-deriving newtype instance Applicative f => Applicative (Joker f a)
-deriving via Compose (p a) f instance
-  (Profunctor p, Applicative (p a), Applicative f)
-    => Applicative (WrappedPafb f p a)
-deriving via Compose (p a) f instance
-  (Profunctor p, Alternative (p a), Applicative f)
-    => Alternative (WrappedPafb f p a)
-instance (Closed p, Distributive f)
-  => Closed (WrappedPafb f p) where
-    closed (WrapPafb p) = WrapPafb (rmap distribute (closed p))
-deriving via (Ap.WrappedArrow p a) instance Arrow p
-  => Functor (Pro.WrappedArrow p a)
-deriving via (Ap.WrappedArrow p a) instance Arrow p
-  => Applicative (Pro.WrappedArrow p a)
-deriving via (Pro.WrappedArrow p) instance Arrow p
-  => Profunctor (Ap.WrappedArrow p)
-instance (Monoidal p, Applicative (q a))
-  => Applicative (Procompose p q a) where
-    pure b = Procompose (pure b) (pure b)
-    Procompose wb aw <*> Procompose vb av = Procompose
-      (dimap2 fst snd ($) wb vb)
-      (liftA2 (,) aw av)
-instance (Monoidal p, Monoidal q)
-  => Applicative (Product p q a) where
-    pure b = Pair (pure b) (pure b)
-    Pair x0 y0 <*> Pair x1 y1 = Pair (x0 <*> x1) (y0 <*> y1)
-instance (Functor f, Functor (p a)) => Functor (Cayley f p a) where
-  fmap f (Cayley x) = Cayley (fmap (fmap f) x)
-instance (Applicative f, Applicative (p a)) => Applicative (Cayley f p a) where
-  pure b = Cayley (pure (pure b))
-  Cayley x <*> Cayley y = Cayley ((<*>) <$> x <*> y)
-instance (Profunctor p, Applicative (p a))
-  => Applicative (Yoneda p a) where
-    pure = proreturn . pure
-    ab <*> cd = proreturn (proextract ab <*> proextract cd)
-instance (Profunctor p, Applicative (p a))
-  => Applicative (Coyoneda p a) where
-    pure = proreturn . pure
-    ab <*> cd = proreturn (proextract ab <*> proextract cd)
-instance (Profunctor p, Alternative (p a))
-  => Alternative (Yoneda p a) where
-    empty = proreturn empty
-    ab <|> cd = proreturn (proextract ab <|> proextract cd)
-    many = proreturn . many . proextract
-instance (Profunctor p, Alternative (p a))
-  => Alternative (Coyoneda p a) where
-    empty = proreturn empty
-    ab <|> cd = proreturn (proextract ab <|> proextract cd)
-    many = proreturn . many . proextract
